Move artichoke-frontend to be the primary workspace crate
Move artichoke-frontend/src to the root src directory in the artichoke
repo.

This change enables installing artichoke and its binaries with cargo install.
This change follows from GH-578 which reserved the artichoke crate on
crates.io.

This change re-exports artichoke-backend from the new artichoke crate.

This PR removes the fixtures and benches directories from artichoke-frontend.

This adds much more detail to the root cargo manifest, preserving some pieces
from artichoke-frontend.

This PR adds installation instructions to the README. This PR adds usage
instructions for airb and artichoke binaries.
