+++ Cargo.toml
@@ -1,10 +1,32 @@
+[package]
+name = "artichoke"
+version = "0.1.0-pre.0"
+authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
+license = "MIT"
+edition = "2018"
+readme = "README.md"
+repository = "https://github.com/artichoke/artichoke"
+documentation = "https://artichoke.github.io/artichoke/artichoke/"
+homepage = "https://github.com/artichoke/artichoke"
+description = """
+artichoke is a Ruby 2.6.3 implementation. Artichoke supports embedding and
+conditional compilation of the VM and standard library. Artichoke can be used to
+distribute Ruby applications as single-binary artifacts. Artichoke builds on
+Windows, macOS, and Linux, as well as WebAssembly via Emscripten.
+"""
+keywords = ["artichoke", "mri", "cruby", "ruby"]
+categories = ["command-line-utilities"]
+
+[dependencies]
+ansi_term = "0.11" # cannot upgrade until clap 3 is released
+rustyline = { version = "6", default-features = false }
+structopt = "0.3.10"
+
+[dependencies.artichoke-backend]
+path = "artichoke-backend"
+
 [workspace]
-members = [
-  "artichoke-backend",
-  "artichoke-core",
-  "artichoke-frontend",
-  "spec-runner",
-]
+members = ["artichoke-backend", "artichoke-core", "spec-runner"]
 
 [profile.release]
 codegen-units = 1
